refactor: Update namespacing for custom extensions
