--- phpstan/extension.neon
@@ -1,5 +1,0 @@
-services:
-  -
-    class: Give\PHPStan\Reflection\LogMethodsReflectionExtension
-    tags:
-      - phpstan.broker.methodsClassReflectionExtension
